imp: Switched u64 to i32 for ValidExpiryPeriod to match error code type of SslExpiration

+++ src/products/expected.rs
@@ -47,7 +47,7 @@ pub enum DomainExpectation {
 
     /// Domain expiry minimum period in days
     #[fail(display = "Passed ValidExpiryPeriod: '{}'", _0)]
-    ValidExpiryPeriod (u64),
+    ValidExpiryPeriod (i32),
 
 }
